app/requirements.txt
customtkinter
Pillow
secure-smtplib
